test(mempool): port #971 e2e mempool tests; drop older duplicates
The soft-finality branch carried an earlier iteration of the mempool
e2e tests folded into tests/tx_batching_test.go. Add the canonical #971
suite in tests/tx_mempool_test.go (adds TTLExpiry and InFlightNonce
rejection cases, uses its own setupTxMemPoolGatewayNode helper) and
remove the now-duplicated Test_TxMemPool_* functions and the redundant
fundEOA helper from tx_batching_test.go, along with the imports only
those tests used.
